refactor(all): 반복문 개선 및 하드코딩 제거
