+++ src/main/java/racingcar/model/Cars.java
@@ -27,7 +27,6 @@ public class Cars {
     public String getWinners() {
         StringJoiner result = new StringJoiner(", ");
         int maxValue = cars.stream().mapToInt(car -> car.getPosition()).max().getAsInt();
-        System.out.println("MAX VALUE: " + maxValue);
         List<Car> winners = cars.stream().filter(car -> car.getPosition() == maxValue).collect(Collectors.toList());
         winners.stream().map(car -> car.getName()).forEach(result::add);
         return result.toString();
